+++ packages/terra-core-docs/src/terra-dev-site/doc/icon/example/IconThemeable.jsx
@@ -2030,8 +2030,7 @@ const IconAll = () => (
         </td>
       </tr>
       <tr style={{ backgroundColor: '#EEEEEE' }}>
-<<<<<<< HEAD
-        <td style={cellStyle}><IconVolumeSetDefault height='2em' width='2em' /></td>
+        <td style={cellStyle}><IconVolumeSetDefault a11yLabel="icon" height='2em' width='2em' /></td>
         <td style={cellStyle}>volumeSet_default</td>
         <td style={{ fontWeight: 'bold', padding: '0.5rem' }}>
           <code>
@@ -2040,7 +2039,7 @@ const IconAll = () => (
         </td>
       </tr>
       <tr style={{ backgroundColor: '#EEEEEE' }}>
-        <td style={cellStyle}><IconVolumeSetDecrease height='2em' width='2em' /></td>
+        <td style={cellStyle}><IconVolumeSetDecrease a11yLabel="icon" height='2em' width='2em' /></td>
         <td style={cellStyle}>volumeSet_decrease</td>
         <td style={{ fontWeight: 'bold', padding: '0.5rem' }}>
           <code>
@@ -2058,7 +2057,7 @@ const IconAll = () => (
         </td>
       </tr>
       <tr style={{ backgroundColor: '#EEEEEE' }}>
-        <td style={cellStyle}><IconVolumeSetMute height='2em' width='2em' /></td>
+        <td style={cellStyle}><IconVolumeSetMute a11yLabel="icon" height='2em' width='2em' /></td>
         <td style={cellStyle}>volumeSet_mute</td>
         <td style={{ fontWeight: 'bold', padding: '0.5rem' }}>
           <code>
@@ -2067,7 +2066,7 @@ const IconAll = () => (
         </td>
       </tr>
       <tr style={{ backgroundColor: '#EEEEEE' }}>
-        <td style={cellStyle}><IconPersonConnection height='2em' width='2em' /></td>
+        <td style={cellStyle}><IconPersonConnection a11yLabel="icon" height='2em' width='2em' /></td>
         <td style={cellStyle}>personConnection</td>
         <td style={{ fontWeight: 'bold', padding: '0.5rem' }}>
           <code>
@@ -2076,7 +2075,7 @@ const IconAll = () => (
         </td>
       </tr>
       <tr style={{ backgroundColor: '#EEEEEE' }}>
-        <td style={cellStyle}><IconTrendingUp height='2em' width='2em' /></td>
+        <td style={cellStyle}><IconTrendingUp a11yLabel="icon" height='2em' width='2em' /></td>
         <td style={cellStyle}>trendingUp</td>
         <td style={{ fontWeight: 'bold', padding: '0.5rem' }}>
           <code>
@@ -2085,7 +2084,7 @@ const IconAll = () => (
         </td>
       </tr>
       <tr style={{ backgroundColor: '#EEEEEE' }}>
-        <td style={cellStyle}><IconTrendingDown height='2em' width='2em' /></td>
+        <td style={cellStyle}><IconTrendingDown a11yLabel="icon" height='2em' width='2em' /></td>
         <td style={cellStyle}>trendingDown</td>
         <td style={{ fontWeight: 'bold', padding: '0.5rem' }}>
           <code>
@@ -2094,10 +2093,7 @@ const IconAll = () => (
         </td>
       </tr>
       <tr style={{ backgroundColor: '#EEEEEE' }}>
-        <td style={cellStyle}><IconCircleSymbol height='2em' width='2em' /></td>
-=======
         <td style={cellStyle}><IconCircleSymbol a11yLabel="icon" height='2em' width='2em' /></td>
->>>>>>> [terra-icon] accessibility upgrade (#3598)
         <td style={cellStyle}>circle_symbol</td>
         <td style={{ fontWeight: 'bold', padding: '0.5rem' }}>
           <code>