+++ packages/terra-icon/src/IconBaseDec.jsx
@@ -50,6 +50,7 @@ const IconBase = ({
   children,
   height,
   width,
+  title,
   focusable,
   ...customProps
 }) => {